feat: revised nc-config
