+++ packages/nocodb/src/index.ts
@@ -1,6 +1,5 @@
 import Noco from './Noco';
-import NcConfigFactory from './utils/NcConfigFactory';
 
 export default Noco;
 
-export { Noco, NcConfigFactory };
+export { Noco };
