+++ packages/nocodb/src/services/projects.service.ts
@@ -11,7 +11,7 @@ import syncMigration from '../helpers/syncMigration';
 import { Project, ProjectUser } from '../models';
 import Noco from '../Noco';
 import extractRolesObj from '../utils/extractRolesObj';
-import NcConfigFactory from '../utils/NcConfigFactory';
+import { getToolDir } from '../utils/nc-config';
 import type { ProjectUpdateReqType } from 'nocodb-sdk';
 import type { ProjectReqType } from 'nocodb-sdk';
 
@@ -96,7 +96,7 @@ export class ProjectsService {
         // if env variable NC_MINIMAL_DBS is set, then create a SQLite file/connection for each project
         // each file will be named as nc_<random_id>.db
         const fs = require('fs');
-        const toolDir = NcConfigFactory.getToolDir();
+        const toolDir = getToolDir();
         const nanoidv2 = customAlphabet(
           '1234567890abcdefghijklmnopqrstuvwxyz',
           14,
