+++ packages/nocodb/src/utils/common/NcConnectionMgr.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import { promisify } from 'util';
 import SqlClientFactory from '../../db/sql-client/lib/SqlClientFactory';
 import { XKnex } from '../../db/CustomKnex';
-import { defaultConnectionConfig } from '../NcConfigFactory';
+import { defaultConnectionConfig } from '../nc-config';
 // import type { NcConfig } from '../../../interface/config';
 import type { Knex } from 'knex';
 // import type NcMetaIO from '../../meta/NcMetaIO';
